requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
supabase
celery[redis]
redis
